requirements.txt
openai
praw
python-dotenv
tqdm
pydantic
